+++ pallets/staking/src/lib.rs
@@ -151,7 +151,7 @@ pub mod pallet {
 
     #[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, TypeInfo, Default)]
     pub struct ReshareInfo<BlockNumber> {
-        pub new_signer: Vec<u8>,
+        pub new_signers: Vec<Vec<u8>>,
         pub block_number: BlockNumber,
     }
 
@@ -189,17 +189,6 @@ pub mod pallet {
     #[pallet::getter(fn threshold_to_stash)]
     pub type ThresholdToStash<T: Config> =
         StorageMap<_, Blake2_128Concat, T::AccountId, T::ValidatorId, OptionQuery>;
-
-    /// Tracks wether the validator's kvdb is synced using a stash key as an identifier
-    #[pallet::storage]
-    #[pallet::getter(fn is_validator_synced)]
-    pub type IsValidatorSynced<T: Config> = StorageMap<
-        _,
-        Blake2_128Concat,
-        <T as pallet_session::Config>::ValidatorId,
-        bool,
-        ValueQuery,
-    >;
 
     #[derive(
         Encode, Decode, Clone, PartialEq, Eq, RuntimeDebug, TypeInfo, MaxEncodedLen, Default,
@@ -296,7 +285,6 @@ pub mod pallet {
 
                 ThresholdServers::<T>::insert(validator_stash, server_info.clone());
                 ThresholdToStash::<T>::insert(&server_info.tss_account, validator_stash);
-                IsValidatorSynced::<T>::insert(validator_stash, true);
             }
 
             let refresh_info = RefreshInfo {
@@ -307,14 +295,17 @@ pub mod pallet {
             // mocks a signer rotation for tss new_reshare tests
             if self.mock_signer_rotate.0 {
                 let next_signers = &mut self.mock_signer_rotate.1.clone();
-                next_signers.push(self.mock_signer_rotate.2[0].clone());
+                let mut new_signers = vec![];
+                for new_signer in self.mock_signer_rotate.2.clone() {
+                    next_signers.push(new_signer.clone());
+                    new_signers.push(new_signer.encode())
+                }
                 let next_signers = next_signers.to_vec();
                 NextSigners::<T>::put(NextSignerInfo { next_signers, confirmations: vec![] });
-
                 ReshareData::<T>::put(ReshareInfo {
                     // To give enough time for test_reshare setup
                     block_number: TEST_RESHARE_BLOCK_NUMBER.into(),
-                    new_signer: self.mock_signer_rotate.clone().2[0].encode(),
+                    new_signers,
                 })
             }
         }
@@ -518,7 +509,6 @@ pub mod pallet {
                 let server_info =
                     ThresholdServers::<T>::take(&validator_id).ok_or(Error::<T>::NoThresholdKey)?;
                 ThresholdToStash::<T>::remove(&server_info.tss_account);
-                IsValidatorSynced::<T>::remove(&validator_id);
                 Self::deposit_event(Event::NodeInfoRemoved(controller));
             }
             Ok(Some(<T as Config>::WeightInfo::withdraw_unbonded(
@@ -607,19 +597,7 @@ pub mod pallet {
             Ok(())
         }
 
-        /// Let a validator declare if their kvdb is synced or not synced
-        /// `synced`: State of validator's kvdb
         #[pallet::call_index(6)]
-        #[pallet::weight(<T as Config>::WeightInfo::declare_synced())]
-        pub fn declare_synced(origin: OriginFor<T>, synced: bool) -> DispatchResult {
-            let who = ensure_signed(origin.clone())?;
-            let stash = Self::threshold_to_stash(who).ok_or(Error::<T>::NoThresholdKey)?;
-            IsValidatorSynced::<T>::insert(&stash, synced);
-            Self::deposit_event(Event::ValidatorSyncStatus(stash, synced));
-            Ok(())
-        }
-
-        #[pallet::call_index(7)]
         #[pallet::weight(({
             <T as Config>::WeightInfo::confirm_key_reshare_confirmed(MAX_SIGNERS as u32)
             .max(<T as Config>::WeightInfo::confirm_key_reshare_completed())
@@ -743,15 +721,46 @@ pub mod pallet {
                 return Ok(weight);
             }
 
-            let mut new_signer = vec![];
+            let mut new_signers: Vec<Vec<u8>> = vec![];
             let mut count = 0u32;
+            let mut remove_indicies_len = 0;
+            // removes first signer and pushes new signer to back if total signers not increased
+            if current_signers_length >= signers_info.total_signers as usize {
+                let mut remove_indicies = vec![];
+                // Finds signers that are no longer validators to remove
+                for (i, current_signer) in current_signers.clone().into_iter().enumerate() {
+                    if !validators.contains(&current_signer) {
+                        remove_indicies.push(i);
+                    }
+                }
+                if remove_indicies.is_empty() {
+                    current_signers.remove(0);
+                } else {
+                    remove_indicies_len = remove_indicies.len();
+                    // reverses vec so as signers removed it does not change location
+                    let remove_indicies_reversed: Vec<_> = remove_indicies.iter().rev().collect();
+                    // truncated as a current limitation see issue: https://github.com/entropyxyz/entropy-core/issues/1114
+                    let truncated = if remove_indicies_reversed.len()
+                        >= (signers_info.total_signers as usize - signers_info.threshold as usize)
+                    {
+                        remove_indicies_reversed[..(signers_info.total_signers as usize
+                            - signers_info.threshold as usize)]
+                            .to_vec()
+                    } else {
+                        remove_indicies_reversed
+                    };
 
-            if current_signers_length <= signers_info.total_signers as usize {
+                    for remove_index in truncated {
+                        current_signers.remove(*remove_index);
+                    }
+                }
+            }
+
+            while current_signers.len() < signers_info.total_signers as usize {
                 let mut randomness = Self::get_randomness();
                 // grab a current signer to initiate value
                 let mut next_signer_up = &current_signers[0].clone();
                 let mut index;
-
                 // loops to find signer in validator that is not already signer
                 while current_signers.contains(next_signer_up) {
                     index = randomness.next_u32() % validators.len() as u32;
@@ -760,14 +769,8 @@ pub mod pallet {
                 }
 
                 current_signers.push(next_signer_up.clone());
-                new_signer = next_signer_up.encode();
+                new_signers.push(next_signer_up.encode());
             }
-
-            // removes first signer and pushes new signer to back if total signers not increased
-            if current_signers_length >= signers_info.total_signers as usize {
-                current_signers.remove(0);
-            }
-
             NextSigners::<T>::put(NextSignerInfo {
                 next_signers: current_signers.clone(),
                 confirmations: vec![],
@@ -777,7 +780,7 @@ pub mod pallet {
             let current_block_number = <frame_system::Pallet<T>>::block_number();
             let reshare_info = ReshareInfo {
                 block_number: current_block_number + sp_runtime::traits::One::one(),
-                new_signer,
+                new_signers,
             };
 
             ReshareData::<T>::put(reshare_info);
@@ -785,7 +788,12 @@ pub mod pallet {
                 jump_start_details.parent_key_threshold = signers_info.threshold
             });
 
-            weight = <T as Config>::WeightInfo::new_session(current_signers.len() as u32, count);
+            weight = <T as Config>::WeightInfo::new_session(
+                current_signers.len() as u32,
+                count,
+                validators.len() as u32,
+                remove_indicies_len as u32,
+            );
 
             Ok(weight)
         }